+++ src/slskd/Shares/SharedFileCache.cs
@@ -30,7 +30,6 @@ namespace slskd.Shares
     using System.Threading.Tasks;
     using Microsoft.Data.Sqlite;
     using Serilog;
-    using Soulseek;
 
     /// <summary>
     ///     Shared file cache.
@@ -49,13 +48,6 @@ namespace slskd.Shares
         /// <param name="filters">The list of regular expressions used to exclude files or paths from scanning.</param>
         /// <returns>The operation context.</returns>
         Task FillAsync(IEnumerable<Share> shares, IEnumerable<Regex> filters);
-
-        /// <summary>
-        ///     Searches the cache for the specified <paramref name="query"/> and returns the matching files.
-        /// </summary>
-        /// <param name="query">The query for which to search.</param>
-        /// <returns>The matching files.</returns>
-        IEnumerable<File> Search(SearchQuery query);
 
         /// <summary>
         ///     Cancels the currently running fill operation, if one is running.
@@ -361,59 +353,6 @@ namespace slskd.Shares
             {
                 CancellationTokenSource = null;
                 SyncRoot.Release();
-            }
-        }
-
-        /// <summary>
-        ///     Searches the cache for the specified <paramref name="query"/> and returns the matching files.
-        /// </summary>
-        /// <param name="query">The query for which to search.</param>
-        /// <returns>The matching files.</returns>
-        public IEnumerable<File> Search(SearchQuery query)
-        {
-            string Clean(string str) => str.Replace("/", " ")
-                .Replace("\\", " ")
-                .Replace(":", " ")
-                .Replace("\"", " ")
-                .Replace("'", "''");
-
-            var match = string.Join(" AND ", query.Terms.Select(token => $"\"{Clean(token)}\""));
-            var exclusions = string.Join(" OR ", query.Exclusions.Select(exclusion => $"\"{Clean(exclusion)}\""));
-
-            var sql = $"SELECT files.maskedFilename, files.code, files.size, files.extension, files.attributeJson FROM filenames " +
-                "INNER JOIN files ON filenames.maskedFilename = files.maskedFilename " +
-                $"WHERE filenames MATCH '({match}) {(query.Exclusions.Any() ? $"NOT ({exclusions})" : string.Empty)}' " +
-                "ORDER BY filenames.maskedFilename ASC;";
-
-            var results = new List<File>();
-            SqliteDataReader reader = default;
-
-            try
-            {
-                using var conn = GetConnection();
-                using var cmd = new SqliteCommand(sql, conn);
-                reader = cmd.ExecuteReader();
-
-                while (reader.Read())
-                {
-                    var filename = reader.GetString(0);
-                    var code = reader.GetInt32(1);
-                    var size = reader.GetInt64(2);
-                    var extension = reader.GetString(3);
-                    var attributeJson = reader.GetString(4);
-
-                    var attributeList = attributeJson.FromJson<List<FileAttribute>>();
-
-                    var file = new File(code, filename, size, extension, attributeList);
-                    results.Add(file);
-                }
-
-                return results;
-            }
-            catch (Exception ex)
-            {
-                Log.Warning(ex, "Failed to execute shared file query '{Query}': {Message}", query, ex.Message);
-                return Enumerable.Empty<File>();
             }
         }
